fix bugs

+++ conv2d/test_gradient.py
@@ -1,8 +1,8 @@
 import torch
-from conv2d.conv2d_functions import Conv2dFunction
+from conv2d_functions import Conv2dFunction
 from torch.autograd import Variable
 import os
-from conv2d.gradcheck import gradcheck
+from gradcheck import gradcheck
 
 os.environ['CUDA_VISIBLE_DEVICES'] = '4'
 batchsize = 2
@@ -16,10 +16,10 @@ dilation = 2
 out = int((inpu + 2 * pad - dilation * (kernel - 1) - 1) / stri + 1)
 group = 2
 
-inputs = Variable(torch.rand(batchsize, c_in, inpu, inpu).double().cuda(), requires_grad=True)
-weight = Variable(torch.rand(c_out, c_in // group, kernel, kernel).double().cuda(),
+inputs = Variable(torch.rand(batchsize, c_in, inpu, inpu).cuda(), requires_grad=True)
+weight = Variable(torch.rand(c_out, c_in // group, kernel, kernel).cuda(),
                   requires_grad=True)
-bias = Variable(torch.rand(c_out).double().cuda(),
+bias = Variable(torch.rand(c_out).cuda(),
                 requires_grad=True)
 
 print(
